fix: Prevent infinite loops
