+++ tests/infinite_loop/a.yml
@@ -1,0 +1,1 @@
+extend: b.yml
