+++ tests/infinite_loop/b.yml
@@ -1,0 +1,1 @@
+extend: a.yml
