+++ tests/infinite_loop/self.yml
@@ -1,0 +1,1 @@
+extend: self.yml
