+++ tests/infinite_loop/stack_size_0.yml
@@ -1,0 +1,1 @@
+extend: stack_size_1.yml
