+++ tests/infinite_loop/stack_size_1.yml
@@ -1,0 +1,1 @@
+extend: stack_size_2.yml
