+++ tests/infinite_loop/stack_size_3.yml
@@ -1,0 +1,1 @@
+extend: stack_size_4.yml
