+++ tests/infinite_loop/stack_size_7.yml
@@ -1,0 +1,1 @@
+extend: stack_size_8.yml
