+++ tests/infinite_loop/stack_size_9.yml
@@ -1,0 +1,1 @@
+from: alpine
